feat(index): move blog post to index
✓ remove guides page
✓ adjust footer for mobile
✓ add redirects for blog posts
