+++ netlify.toml
@@ -9,3 +9,11 @@
 
 [[plugins]]
   package = "netlify-plugin-gatsby-cache"
+
+[[redirects]]
+  from = "/guides/2020-05-17-collaborative-ui-development"
+  to = "/2020-05-17-collaborative-ui-development"
+
+[[redirects]]
+  from = "/guides/2020-05-02-continuous-integration-with-netlify-build-plugins"
+  to = "/2020-05-02-continuous-integration-with-netlify-build-plugins"
