+++ src/components/layout.tsx
@@ -24,17 +24,10 @@ const GlobalStyle = createGlobalStyle`
 export const Layout: React.FC<LayoutProps> = ({ children, size }) => {
   const theme = useTheme();
 
-  const { isMobileNavigationOpen, setIsMobileNavigationOpen } = useContext(
+  const { isMobileNavigationOpen } = useContext(
     MobileNavigationContext
   );
   const isNarrow = size === 'narrow';
-
-  const handleMobileNavigation = () => {
-    if (isMobileNavigationOpen) {
-      return setIsMobileNavigationOpen(false);
-    }
-    return;
-  };
 
   return (
     <>
@@ -52,15 +45,6 @@ export const Layout: React.FC<LayoutProps> = ({ children, size }) => {
         height={isMobileNavigationOpen ? '100vh' : 'auto'}
       >
         <Navigation>
-          <Link
-            variant="tertiary"
-            component={GatsbyLink}
-            onClick={handleMobileNavigation}
-            to="/guides"
-            marginBottom={[theme.space.l, 0]}
-          >
-            Guides
-          </Link>
           <SocialLink href="https://twitter.com/_yetanotherblog">
             <Twitter />
           </SocialLink>
